move pull command to pull.mjs, add missing import
