+++ packages/usdk/lib/create.mjs
@@ -595,70 +595,3 @@ export const edit = async (args, opts) => {
   };
   await _updateFiles();
 };
-export const pull = async (args, opts) => {
-  const agentId = args._[0] ?? '';
-  const dstDir = args._[1] ?? cwd;
-  const force = !!args.force;
-  const forceNoConfirm = !!args.forceNoConfirm;
-  const noInstall = !!args.noInstall;
-  const events = args.events ?? null;
-  // opts
-  const jwt = opts.jwt;
-  if (!jwt) {
-    throw new Error('You must be logged in to pull.');
-  }
-
-  const userId = jwt && (await getUserIdForJwt(jwt));
-  if (userId) {
-    // clean the old directory
-    await cleanDir(dstDir, {
-      force,
-      forceNoConfirm,
-    });
-
-    // download the source
-    console.log(pc.italic('Downloading source...'));
-    const u = `https://${aiProxyHost}/agents/${agentId}/source`;
-    try {
-      const req = await fetch(u, {
-        headers: {
-          Authorization: `Bearer ${jwt}`,
-        },
-      });
-      if (req.ok) {
-        const zipBuffer = await req.arrayBuffer();
-        // console.log('downloaded source', zipBuffer.byteLength);
-
-        // extract the source
-        console.log(pc.italic('Extracting zip...'));
-        await extractZip(zipBuffer, dstDir);
-
-        events && events.dispatchEvent(new MessageEvent('finalize', {
-          data: {
-            agentJson,
-          },
-        }));
-
-        console.log(pc.italic('Installing dependencies...'));
-        try {
-          if (!noInstall) {
-            await npmInstall(dstDir);
-          }
-        } catch (err) {
-          console.warn('npm install failed:', err.stack);
-          process.exit(1);
-        }
-      } else {
-        const text = await req.text();
-        console.warn('pull request error', text);
-        process.exit(1);
-      }
-    } catch (err) {
-      console.warn('pull request failed', err);
-      process.exit(1);
-    }
-  } else {
-    console.log('not logged in');
-    process.exit(1);
-  }
-};
